Prefer the wildcard registration on Windows
Registers all file extensions with the extended context menu upon integration

+++ Athena.Core/Desktop/Windows/RegistryEntry.cs
@@ -99,14 +99,14 @@ public class RegistryEntry
         key.Close();
         
         // HKCU\Software\Classes\*
-        if (fullDeletion)
+        if (fullDeletion && fileExtension != "*")
         {
             Registry.CurrentUser.DeleteSubKeyTree(keyPath);
             return;
         }
         
         // HKCU\Software\Classes\*\shell\OpenWithAthena
-        Registry.CurrentUser.DeleteSubKeyTree($@"{keyPath}\shell\OpenWithAthena");
+        Registry.CurrentUser.DeleteSubKeyTree($@"{keyPath}\shell\OpenWithAthena", false);
     }
     
     private static void RemoveProtocolFromContextMenu(string protocol)
